backend/requirements.txt
alembic==1.16.4
bcrypt==4.0.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Admin==1.6.1
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Unidecode==1.4.0
Werkzeug==3.1.3
WTForms==3.2.1
